+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Teknoo Software - East Foundation - Change Log
 
-## [9.2.0] - 2026-03-26
+## [9.2.0] - 2026-03-28
 ### Stable Release
 - Add `Teknoo\East\Foundation\Normalizer\Object\AutoTrait` and attributes
-  - `Teknoo\East\Foundation\Normalizer\Object\Group`
+  - `Teknoo\East\Foundation\Normalizer\Object\Normalize`
   - `Teknoo\East\Foundation\Normalizer\Object\ClassGroup`
   to implements easily `Teknoo\East\Foundation\Normalizer\EastNormalizerInterface` by defining groups with previous
   attributes and avoid fill an array to pass to 
     `Teknoo\East\Foundation\Normalizer\Object\GroupTrait::setGroupsConfiguration()`
+  - `Normalize` attribute allow also to define an alias instead of the property name and a callable to load and fetch 
+    the value to export. The loader can be a `static closure` in PHP 8.5, an `array callable`, a method name of the 
+    current object, or the joker `@lazy` and `AutoTrait` will create a static closure to return the value only after 
+    filtering.
 
 ## [9.1.1] - 2026-03-11
 ### Stable Release